receivinglisttotals and modal ts
